Dependencies: Move YAML reader and writer code to swift-dependency-tool
