+++ lib/AST/FineGrainedDependencies.cpp
@@ -28,7 +28,6 @@
 #include "llvm/ADT/SmallVector.h"
 #include "llvm/Support/FileSystem.h"
 #include "llvm/Support/Path.h"
-#include "llvm/Support/YAMLParser.h"
 
 
 // This file holds the definitions for the fine-grained dependency system
@@ -371,88 +370,3 @@ void SourceFileDepGraph::emitDotFile(StringRef outputPath,
     return false;
   });
 }
-
-//==============================================================================
-// MARK: SourceFileDepGraph YAML reading & writing
-//==============================================================================
-
-namespace llvm {
-namespace yaml {
-// This introduces a redefinition for Linux.
-#if !(defined(__linux__) || defined(_WIN64))
-void ScalarTraits<size_t>::output(const size_t &Val, void *, raw_ostream &out) {
-  out << Val;
-}
-
-StringRef ScalarTraits<size_t>::input(StringRef scalar, void *ctxt,
-                                      size_t &value) {
-  return scalar.getAsInteger(10, value) ? "could not parse size_t" : "";
-}
-#endif
-
-void ScalarEnumerationTraits<swift::fine_grained_dependencies::NodeKind>::
-    enumeration(IO &io, swift::fine_grained_dependencies::NodeKind &value) {
-  using NodeKind = swift::fine_grained_dependencies::NodeKind;
-  io.enumCase(value, "topLevel", NodeKind::topLevel);
-  io.enumCase(value, "nominal", NodeKind::nominal);
-  io.enumCase(value, "potentialMember", NodeKind::potentialMember);
-  io.enumCase(value, "member", NodeKind::member);
-  io.enumCase(value, "dynamicLookup", NodeKind::dynamicLookup);
-  io.enumCase(value, "externalDepend", NodeKind::externalDepend);
-  io.enumCase(value, "sourceFileProvide", NodeKind::sourceFileProvide);
-}
-
-void ScalarEnumerationTraits<DeclAspect>::enumeration(
-    IO &io, swift::fine_grained_dependencies::DeclAspect &value) {
-  using DeclAspect = swift::fine_grained_dependencies::DeclAspect;
-  io.enumCase(value, "interface", DeclAspect::interface);
-  io.enumCase(value, "implementation", DeclAspect::implementation);
-}
-
-void MappingTraits<DependencyKey>::mapping(
-    IO &io, swift::fine_grained_dependencies::DependencyKey &key) {
-  io.mapRequired("kind", key.kind);
-  io.mapRequired("aspect", key.aspect);
-  io.mapRequired("context", key.context);
-  io.mapRequired("name", key.name);
-}
-
-void MappingTraits<DepGraphNode>::mapping(
-    IO &io, swift::fine_grained_dependencies::DepGraphNode &node) {
-  io.mapRequired("key", node.key);
-  io.mapOptional("fingerprint", node.fingerprint);
-}
-
-void MappingContextTraits<SourceFileDepGraphNode, SourceFileDepGraph>::mapping(
-    IO &io, SourceFileDepGraphNode &node, SourceFileDepGraph &g) {
-  MappingTraits<DepGraphNode>::mapping(io, node);
-  io.mapRequired("sequenceNumber", node.sequenceNumber);
-  std::vector<size_t> defsIDependUponVec(node.defsIDependUpon.begin(),
-                                         node.defsIDependUpon.end());
-  io.mapRequired("defsIDependUpon", defsIDependUponVec);
-  io.mapRequired("isProvides", node.isProvides);
-  if (!io.outputting()) {
-    for (size_t u : defsIDependUponVec)
-      node.defsIDependUpon.insert(u);
-  }
-  assert(g.getNode(node.sequenceNumber) && "Bad sequence number");
-}
-
-size_t SequenceTraits<std::vector<SourceFileDepGraphNode *>>::size(
-    IO &, std::vector<SourceFileDepGraphNode *> &vec) {
-  return vec.size();
-}
-
-SourceFileDepGraphNode &
-SequenceTraits<std::vector<SourceFileDepGraphNode *>>::element(
-    IO &, std::vector<SourceFileDepGraphNode *> &vec, size_t index) {
-  while (vec.size() <= index)
-    vec.push_back(new SourceFileDepGraphNode());
-  return *vec[index];
-}
-
-void MappingTraits<SourceFileDepGraph>::mapping(IO &io, SourceFileDepGraph &g) {
-  io.mapRequired("allNodes", g.allNodes, g);
-}
-} // namespace yaml
-} // namespace llvm
